Rename 'redwood-compose-testing' to 'redwood-testing'
This has more responsibility now and deserves the wider name.
