--- redwood-compose-testing/gradle.properties
@@ -1,2 +1,0 @@
-POM_ARTIFACT_ID=redwood-compose-testing
-POM_NAME=Redwood Compose Testing
