+++ redwood-testing/gradle.properties
@@ -1,0 +1,2 @@
+POM_ARTIFACT_ID=redwood-testing
+POM_NAME=Redwood Testing
