+++ redwood-testing/src/commonMain/kotlin/app/cash/redwood/testing/TestRedwoodComposition.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.redwood.compose.testing
+package app.cash.redwood.testing
 
 import androidx.compose.runtime.BroadcastFrameClock
 import androidx.compose.runtime.Composable
