+++ redwood-testing/src/commonMain/kotlin/app/cash/redwood/testing/toChangeList.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.redwood.compose.testing
+package app.cash.redwood.testing
 
 import app.cash.redwood.protocol.SnapshotChangeList
 import app.cash.redwood.protocol.compose.ProtocolBridge
