+++ samples/emoji-search/presenter/build.gradle
@@ -20,7 +20,7 @@ kotlin {
         implementation libs.kotlin.test
         implementation libs.assertk
         implementation libs.kotlinx.coroutines.test
-        implementation projects.redwoodComposeTesting
+        implementation projects.redwoodTesting
         implementation projects.samples.emojiSearch.schema.testing
       }
     }
