+++ settings.gradle
@@ -40,7 +40,6 @@ enableFeaturePreview('TYPESAFE_PROJECT_ACCESSORS')
 rootProject.name = 'redwood'
 
 include ':redwood-compose'
-include ':redwood-compose-testing'
 include ':redwood-flexbox'
 include ':redwood-gradle-plugin'
 include ':redwood-layout-api'
@@ -66,6 +65,7 @@ include ':redwood-protocol-compose'
 include ':redwood-protocol-widget'
 include ':redwood-runtime'
 include ':redwood-schema'
+include ':redwood-testing'
 include ':redwood-tooling-codegen'
 include ':redwood-tooling-lint'
 include ':redwood-tooling-schema'
